requirements.txt
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
watchdog==2.1.9
pystray==0.19.4
Pillow==9.5.0
flask==2.3.3
flask-cors==4.0.0